# docs/openspec/registry.openapi.yaml
openapi: 3.0.3
info:
  title: Virbius Control Plane API (MVP)
  version: 1.3.0-mvp
  description: |
    **virbius-control** 单 Java 进程（Registry + Admin 合并）。

    - **规范面**：`/api/v1/admin/tenants/{tenantId}/...`，响应 `{ code, message, data, timestamp }`（`code=0` 成功）。
    - **兼容面**：`/api/v1/tenants/{tenantId}/...`（Legacy，裸 JSON），逐步废弃。
    - 规则真源：`rule_history`；`runtime` 见 DESIGN §8.5.0。
    - **规则级 enforce / canary**（管侧真拦 + 云 PolicyMerge）：见 [rule-level-enforce.md](./rule-level-enforce.md)。
    - PoC 发布为 **同步 HTTP 200**（非 202 异步）。
    - Bundle YAML 导入（`draft/import`）**未实现**，Post-MVP。
servers:
  - url: http://virbius-control:8080
    description: PoC（Registry + Admin UI 同端口）
...
paths:
...
  /api/v1/tenants/{tenantId}/bundles:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 列出 Bundle"
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: 裸 JSON 数组
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BundleSummary'
  /api/v1/tenants/{tenantId}/bundles/{bundleId}/versions/{version}:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 获取 Bundle 版本"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/metadata:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 获取 Bundle 元数据（无 PUT context-bindings）"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleMetadata'
  /api/v1/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/publish:
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 发布 Bundle（同步 200）"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: 发布完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishStatus'
  /api/v1/tenants/{tenantId}/access-lists:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 查询全部名单"
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/{dimension}/{polarity}:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 查询单维名单（无 PUT 整表替换）"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/{dimension}/{polarity}/entries:
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 追加名单条目"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessListEntriesRequest'
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/{dimension}/{polarity}/entries/{value}:
    delete:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 删除名单条目"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
        - name: value
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/sync-rules:
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 同步名单规则"
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/push-engine:
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 推送 engine"
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/rules:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 列出规则"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - name: layer
          in: query
          schema:
            type: string
            enum: [edge, gateway, cloud]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RulePointer'
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 创建或更新规则"
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertRuleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleRevision'
  /api/v1/tenants/{tenantId}/rules/{ruleId}:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 获取规则"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleRevision'
  /api/v1/tenants/{tenantId}/rules/{ruleId}/runtime:
    patch:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 规则级放量（enforce_mode / canary_percent）"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/ruleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuntimeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleRevision'
components:
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        example: default
    bundleId:
      name: bundleId
      in: path
      required: true
      schema:
        type: string
        example: poc-default
    version:
      name: version
      in: path
      required: true
      schema:
        type: string
        example: "0.1.0"
    ruleId:
      name: ruleId
      in: path
      required: true
      schema:
        type: string
...
    dimension:
      name: dimension
      in: path
      required: true
      schema:
        type: string
        enum: [keyword, user_id, device_id, ip_cidr, var]
    polarity:
      name: polarity
      in: path
      required: true
      schema:
        type: string
        enum: [deny, allow]
...
  schemas:
...
    BundleSummary:
      type: object
      properties:
        bundle_id:
          type: string
        tenant_id:
          type: string
        latest_version:
          type: string
        status:
          type: string
          enum: [draft, validating, eval_running, compiling, syncing, active, failed]
...
    BundleMetadata:
      type: object
      properties:
        tenant_id:
          type: string
        bundle_id:
          type: string
        version:
          type: string
        status:
          type: string
        metadata:
          type: object
          additionalProperties: true
        context_bindings:
          $ref: '#/components/schemas/ContextBindingsBlock'
        context_vars:
          type: array
          items:
            $ref: '#/components/schemas/ContextVarBindingRow'
...
    ContextVarBindingRow:
      type: object
      required: [logical, from]
      properties:
        logical:
          type: string
          pattern: '^[a-z][a-z0-9_]*$'
        from:
          type: string
          enum: [query, header, subject, network]
        name:
          type: string
          description: from=query|header 时 HTTP 名；缺省=logical
        field:
          type: string
          description: from=subject|network 时字段名；缺省=logical
    ContextBindingsBlock:
      type: object
      properties:
        version:
          type: integer
          example: 1
        vars:
          type: object
          additionalProperties:
            type: object
            properties:
              from:
                type: string
                enum: [query, header, subject, network]
              name:
                type: string
              field:
                type: string
    AccessListEntriesRequest:
      type: object
      properties:
        value:
          type: string
        values:
          type: array
          items:
            type: string
    PublishStatus:
      type: object
      properties:
        bundle_id:
          type: string
        bundle_version:
          type: string
        status:
          type: string
        publish_id:
          type: string
        sync_ack:
          type: object
          additionalProperties: true
    RulePointer:
      type: object
      properties:
        rule_id:
          type: string
        current_revision:
          type: integer
        layer:
          type: string
        runtime:
          type: string
          enum: [lua-dsl, lua, prompt, groovy, list_match, cumulative]
        reason_code:
          type: string
        intent_action:
          type: string
          enum: [allow, deny, captcha, review]
          default: deny
          description: 规则命中后的希望动作；见 rule-level-enforce.md
        enforce_mode:
          type: string
          enum: [dry_run, canary, full]
          description: 规则级执行模式；见 rule-level-enforce.md
        canary_percent:
          type: integer
          minimum: 1
          maximum: 100
          nullable: true
          description: enforce_mode=canary 时必填；full/dry_run 时应省略或 null
        rule_status:
          type: string
          enum: [draft, active, disabled]
          description: |
            draft=配置中不进执行面；active=参与匹配；disabled=停用不可编辑。
            流转：draft→active|disabled；active→disabled；disabled→draft。
...
    UpsertRuleRequest:
      type: object
      required: [rule_id, layer, runtime, reason_code, body]
      properties:
        rule_id:
          type: string
        bundle_id:
          type: string
        layer:
          type: string
          enum: [edge, gateway, cloud]
        runtime:
          type: string
          enum: [lua-dsl, lua, prompt, groovy, list_match, cumulative]
          description: 详见 DESIGN.md §8.5.0
        reason_code:
          type: string
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
          default: 100
        intent_action:
          type: string
          enum: [allow, deny, captcha, review]
          default: deny
          description: 规则命中后的希望动作
        scope:
          type: object
        body:
          oneOf:
            - type: string
            - type: object
      description: |
        新建或更新规则正文。**不包含** enforce / canary（默认 dry_run）。
        放量使用 `PATCH .../rules/{ruleId}/runtime`。
    UpdateRuntimeRequest:
      type: object
      required: [enforce_mode]
      properties:
        enforce_mode:
          type: string
          enum: [dry_run, canary, full]
          description: 规则级 enforce；见 rule-level-enforce.md
        canary_percent:
          type: integer
          minimum: 1
          maximum: 100
          nullable: true
          description: enforce_mode=canary 时必填；否则省略或 null
    RuleRevisionMeta:
      type: object
      properties:
        rule_revision:
          type: integer
        enforce_mode:
          type: string
          enum: [dry_run, canary, full]
        canary_percent:
          type: integer
          minimum: 1
          maximum: 100
          nullable: true
        rule_status:
          type: string
          enum: [draft, active, disabled]
          description: |
            draft=配置中不进执行面；active=参与匹配；disabled=停用不可编辑。
            流转：draft→active|disabled；active→disabled；disabled→draft。
        modified_at:
          type: string
          format: date-time
        effective_from:
          type: string
          format: date-time
        effective_to:
          type: string
          format: date-time
          nullable: true
    RuleRevision:
      allOf:
        - $ref: '#/components/schemas/RuleRevisionMeta'
        - type: object
          properties:
            tenant_id:
              type: string
            rule_id:
              type: string
            layer:
              type: string
            runtime:
              type: string
            reason_code:
              type: string
            risk_score:
              type: integer
            intent_action:
              type: string
              enum: [allow, deny, captcha, review]
            scope:
              type: object
            body:
              oneOf:
                - type: string
                - type: object
            body_hash:
              type: string
...